\name{DrawLatLon}
\alias{DrawLatLon}
\title{Add a grid of meridians and latitude lines.}
\description{}
\usage{
DrawLatLon(nx=9, ny=9, labels = TRUE, pretty = TRUE, cex = 1, 
            col = "blue", lty = 2, font = 2, lab.x=2,lab.y=2, ...)
}
\arguments{
  \item{nx,ny}{The (approximate) numbr of lines to be drawn.}
  \item{labels}{If TRUE, the latitude and longitude values are printed.}
  \item{pretty}{If TRUE, "nice" values are used.}
  \item{lab.x,lab.y}{The grid line along which the labels should be written.}
  \item{...}{Graphical parameters (cex=text size, lty=line type) .}
}
\value{
}
\details{
}
\examples{
}
\seealso{\code{DomainPoints}
}
\keyword{file}
