\name{local}
\alias{lambel1972}
\alias{lambel2008}
\alias{DomainGLAMEPSv0}
\alias{DomainGLAMEPSv1}
\alias{DomainGLAMEPSv2}
\alias{DomainGLAMEPSv3}
\alias{DomainGLAMEPSv1.fa}
\alias{DomainGLAMEPSv2.fa}
\alias{DomainGLAMEPSv3.fa}
\alias{DomainBel4}
\alias{DomainBel7}
\alias{DomainTeleOld}
\alias{DomainTele}
\alias{DomainINCA}
\title{Local datasets}
\description{Some domains and projections defined for RMI.}

