\name{utilities}
\alias{i2a}
\alias{matsum}
\title{Simple utilities}
\description{Some simple functions that may be useful. Just simple shortcuts.}
\usage{
i2a(x,n)
matsum(x)
}
\arguments{
  \item{x,n}{integers}
  \item{x}{A matrix or array }
}
\value{\code{i2a} formats x as an integer of the required length with zero padding. \code{matsum} is simply the summary of all the  
}
\keyword{file}
